move transform function into its own module.
