+++ tests/unit/transformer/test_node_transformer.py
@@ -9,8 +9,9 @@ import asttokens
 import pytest
 from mock import MagicMock, patch
 
-from lineapy.transformer.node_transformer import NodeTransformer, transform
+from lineapy.transformer.node_transformer import NodeTransformer
 from lineapy.transformer.source_giver import SourceGiver
+from lineapy.transformer.transform_code import transform
 
 
 def _get_ast_node(code):
@@ -23,7 +24,7 @@ def _get_ast_node(code):
 
 
 @patch(
-    "lineapy.transformer.node_transformer.NodeTransformer",
+    "lineapy.transformer.transform_code.NodeTransformer",
 )
 def test_transform_fn(nt_mock: MagicMock):
     """
